fix(tests): drop float32 chain-id casts in integration canary test
CI compiles the integration-tagged tests; the chain-id fields are int now.

+++ tests/integration/canary_test.go
@@ -527,8 +527,8 @@ func runFusionPlusCanary(t *testing.T, mode fusionPlusMode) {
 	require.NoError(t, err)
 
 	quoteParams := fusionplus.QuoterControllerGetQuoteParamsFixed{
-		SrcChain:        float32(src.chainId),
-		DstChain:        float32(dst.chainId),
+		SrcChain:        int(src.chainId),
+		DstChain:        int(dst.chainId),
 		SrcTokenAddress: strings.ToLower(srcToken.Hex()),
 		DstTokenAddress: strings.ToLower(dstToken.Hex()),
 		Amount:          canaryFusionPlusAmount.String(),
